minor changes in meas_*.cu; int => uint*_t

+++ gpu/meas_energy.cu
@@ -31,17 +31,17 @@ int main(int argc, char* argv[])
   // parser for arg list
   argsparse parser(argc, argv, options, defaults);
 
-  const int L = parser.find<int>("L"),
+  const uint32_t L = parser.find<uint32_t>("L"),
     nInputs = L*L,
-    nHiddens = parser.find<int>("nh"),
-    nChains = parser.find<int>("ns"),
-    nWarmup = parser.find<int>("nwarm"),
-    nMonteCarloSteps = parser.find<int>("nms"),
-    deviceNumber = parser.find<int>("dev"),
-    nIterations =  parser.find<int>("niter"),
-    version = parser.find<int>("ver");
+    nHiddens = parser.find<uint32_t>("nh"),
+    nChains = parser.find<uint32_t>("ns"),
+    nWarmup = parser.find<uint32_t>("nwarm"),
+    nMonteCarloSteps = parser.find<uint32_t>("nms"),
+    deviceNumber = parser.find<uint32_t>("dev"),
+    nIterations =  parser.find<uint32_t>("niter"),
+    version = parser.find<uint32_t>("ver");
   const double h = parser.find<double>("h"), J = parser.find<double>("J");
-  const unsigned long long seed = parser.find<unsigned long long>("seed");
+  const uint64_t seed = parser.find<uint64_t>("seed");
   const std::string path = parser.find<>("path") + "/",
     nistr = std::to_string(nInputs),
     nhstr = std::to_string(nHiddens),
@@ -77,10 +77,10 @@ int main(int argc, char* argv[])
   struct SamplerTraits { using AnsatzType = ComplexFNN<double>; using FloatType = double;};
 
   // block size for the block splitting scheme of parallel Monte-Carlo
-  const unsigned long nBlocks = static_cast<unsigned long>(nIterations)*
-                                static_cast<unsigned long>(nMonteCarloSteps)*
-                                static_cast<unsigned long>(nInputs)*
-                                static_cast<unsigned long>(nChains);
+  const uint64_t nBlocks = static_cast<uint64_t>(nIterations)*
+                           static_cast<uint64_t>(nMonteCarloSteps)*
+                           static_cast<uint64_t>(nInputs)*
+                           static_cast<uint64_t>(nChains);
 
   // Transverse Field Ising Hamiltonian on the square lattice
   using SamplerType = spinhalf::TFISQ<SamplerTraits>;
